requirements.txt
durable_rules
langchain
openai